+++ data/data-samples/sample-data-mongodb-reactive/src/test/java/io/americanexpress/data/book/repository/BookRepositoryIT.java
@@ -1,4 +1,4 @@
-package io.americanexpress.data.book.dao;
+package io.americanexpress.data.book.repository;
 
 import io.americanexpress.data.book.config.BookDataTestConfig;
 import io.americanexpress.data.book.entity.BookEntity;
